Added CLI option to enable and clear the cache 🧣
Both eslint and stylelint support a cache which can be used to only
lint files which have changed. This can speed up the linting process
which is important when lint tasks are run in pre-commit hooks.

Having an option to clear the cache is also important. Although users
could just delete the cache files, doing so via the CLI is easy and
acts as a reminder that linting with the cache enabled may have
different results than running in a deployment pipeline where no cache
is available.